Checking Screen Bounds
This update checks the screen bounds as the player moves. There is a bug
in the upper left of the screen where when both checks happen to hit at
the same time, the player can move off screen.

Need to update animation so that I can get sprite w/h back for better
bounds calculations.