Fix Bounds
Had an issue with player going out of bounds if x was 0 or at max
height boundry. Removed the large checkbounds method and added a
func for each direction that simply returns a bool. then added those
checks in with the key checks. Reduced the lines of code and simplified
the process of checking bounds.